refactor(LottoTicket): 추상 클래스로 분리
